# tests_enablement_2.0/test_1/customer-domain-api/src/main/java/com/bank/customer/model/Error.java
package com.bank.customer.model;

import lombok.Data;
import lombok.Builder;
import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

import java.util.List;
import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class Error {

    private String code;
    private String message;
    private List<Detail> details;
    private UUID correlationId;

    @Data
    @Builder
    @NoArgsConstructor
    @AllArgsConstructor
    public static class Detail {
        private String field;
        private String message;
    }

}